requirements.txt
PyGithub
Flask
requests
aiohttp